Instance rendering with font atlas!
Refactor buffer, texture and shader creation. Implements a C++ class
wrapper for these data.

In vertex buffer, it now includes a way to describe the vertex data
layout. This reduce human error when writing the OpenGL code. Not
completely safe because it reuse the OpenGL enum which is not type safe.
This make it possible to pass in other type of data that is not
intended.