Fixes atlas generation border check
Load in new glyph if there's an unknown code. If load success then we
regenerate the texture atlas. Right now it is regenerating the whole
texture. In future work it is better to continue to and where we left
off if there is space in the texture atlas.

Add image to README.